refactor helpicon button to be generic icon button
