+++ src/components/Navigation/Sidebar.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { NavigationLinks } from "./NavigationLinks.tsx";
 import { SidebarProps } from "../../types/types.ts";
 import FileTable from "../FileHandling/FileTable.tsx";
-import HelpIcon from "../utils/HelpIcon.tsx";
+import IconButton from "../utils/IconButton.tsx";
 import Modal from "../utils/Modal.tsx";
 
 /**
@@ -25,7 +25,12 @@ const Sidebar: React.FC<SidebarProps> = ({
         <div className="z-1 fixed right-0 top-0 h-full w-64 overflow-y-auto bg-secondary p-6 py-20 text-secondary-content">
             <h3 className="mb-2 mt-2 text-xl font-semibold">Sidebar</h3>
 
-            <HelpIcon onClick={toggleModal} />
+            <div
+                className="absolute left-3/4 top-7 mt-14 -translate-x-1/2 transform"
+
+            >
+                <IconButton onClick={toggleModal} icon={"help"} />
+            </div>
 
             <NavigationLinks />
