--- src/components/utils/HelpIcon.tsx
@@ -1,40 +1,0 @@
-import React from "react";
-
-/**
- * interface HelpIconProps
- * @param onClick - Function to handle help icon click
- */
-interface HelpIconProps {
-    onClick: () => void;
-}
-
-/**
- *
- * @param onClick - Function to handle help icon click
- * @returns rendered HelpIcon component
- */
-const HelpIcon: React.FC<HelpIconProps> = ({ onClick }) => {
-    return (
-        <div
-            className="absolute left-3/4 top-7 mt-14 -translate-x-1/2 transform cursor-pointer hover:text-accent"
-            onClick={onClick}
-        >
-            <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth={1.5}
-                stroke="currentColor"
-                className="size-8"
-            >
-                <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M9.879 7.519c1.171-1.025 3.071-1.025 4.242 0 1.172 1.025 1.172 2.687 0 3.712-.203.179-.43.326-.67.442-.745.361-1.45.999-1.45 1.827v.75M21 12a9 9 0 1 1-18 0 9 9 0 0 1 18 0Zm-9 5.25h.008v.008H12v-.008Z"
-                />
-            </svg>
-        </div>
-    );
-};
-
-export default HelpIcon;
